Initial import
This patch adds snap-confine, snapd and their build dependencies.  All the
golang-* packages apart from golang-github-jessevdk-go-flags are new and not
available in Fedora 24 yet. The snapd package uses a hack to start the
snapd.socket and snapd.refresh.timer on installation. It should be possible to
do this with presets but I wasn't successful about that.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>